Refactor of media player unit test
