+++ src/app/containers/MediaPlayer/fixtureData.jsx
@@ -1,0 +1,380 @@
+import React from 'react';
+import { string, shape, bool, arrayOf, object, oneOfType } from 'prop-types';
+import { singleTextBlock } from '#models/blocks';
+import { RequestContextProvider } from '#contexts/RequestContext';
+import { ServiceContextProvider } from '#contexts/ServiceContext';
+import { ToggleContext } from '#contexts/ToggleContext';
+import MediaPlayerContainer from '.';
+
+const captionBlock = {
+  model: {
+    blocks: [
+      {
+        model: {
+          blocks: [singleTextBlock('Caption')],
+        },
+      },
+    ],
+  },
+  type: 'caption',
+};
+
+const imageBlock = {
+  model: {
+    blocks: [
+      {
+        model: {
+          copyrightHolder: 'BBC',
+          height: 1080,
+          locator: 'ichef.test.bbci.co.uk/images/ic/$widthxn/p01k6mtv.jpg',
+          originCode: 'mpv',
+          width: 1920,
+        },
+        type: 'rawImage',
+        id: '48e075e3-2f46-4dd5-badf-432a08f66482',
+        position: [4, 2, 2, 1],
+      },
+      {
+        model: {
+          blocks: [singleTextBlock('Ants')],
+        },
+        type: 'altText',
+      },
+    ],
+  },
+  type: 'image',
+  id: '4c000b11-a141-4983-acec-71fd5473e215',
+  position: [4, 2, 2],
+};
+
+const validAresMediaVideoBlock = {
+  model: {
+    blocks: [
+      {
+        blockId: 'urn:bbc:ares::clip:p01k6msm',
+        model: {
+          advertising: true,
+          embedding: true,
+          format: 'audio_video',
+          id: 'p01k6msm',
+          imageCopyright: 'BBC',
+          imageUrl: 'ichef.test.bbci.co.uk/images/ic/$recipe/p01k6mtv.jpg',
+          subType: 'clip',
+          syndication: {
+            destinations: [],
+          },
+          synopses: {
+            short:
+              'They may be tiny, but us humans could learn a thing or two from ants.',
+          },
+          title: 'Five things ants can teach us about management',
+          versions: [
+            {
+              availableFrom: 1540218932000,
+              availableTerritories: {
+                nonUk: true,
+                uk: true,
+              },
+              duration: 191,
+              durationISO8601: 'PT3M11S',
+              types: ['Original'],
+              versionId: 'p01k6msp',
+              warnings: {
+                long: 'Contains strong language and adult humour.',
+                short: 'Contains strong language and adult humour.',
+              },
+            },
+          ],
+        },
+        type: 'aresMediaMetadata',
+      },
+      imageBlock,
+    ],
+  },
+  type: 'aresMedia',
+};
+
+// Block test data below has been commented as they will be used in the future
+// const validAresMediaAudioBlock = {
+//   model: {
+//     blocks: [
+//       {
+//         blockId: 'urn:bbc:ares::clip:p01m7d07',
+//         model: {
+//           advertising: false,
+//           embedding: false,
+//           format: 'audio',
+//           id: 'p01m7d07',
+//           imageCopyright: 'Getty Images',
+//           imageUrl: 'ichef.test.bbci.co.uk/images/ic/$recipe/p01mt2kt.jpg',
+//           subType: 'clip',
+//           syndication: {
+//             destinations: [],
+//           },
+//           synopses: {
+//             short: 'Some audio from a supermarket checkout in Birmingham',
+//           },
+//           title: 'Birmingham checkout',
+//           versions: [
+//             {
+//               availableFrom: 1555067395000,
+//               availableTerritories: {
+//                 nonUk: true,
+//                 uk: true,
+//               },
+//               duration: 127,
+//               durationISO8601: 'PT2M7S',
+//               types: ['Original'],
+//               versionId: 'p01m7d09',
+//               warnings: {
+//                 long: 'Contains some strong language.',
+//                 short: 'Contains some strong language.',
+//               },
+//             },
+//           ],
+//         },
+//         type: 'aresMediaMetadata',
+//       },
+//       imageBlock,
+//     ],
+//   },
+//   type: 'aresMedia',
+// };
+//
+// const missingAresMediaMetadataBlock = {
+//   model: {
+//     blocks: [imageBlock],
+//   },
+//   type: 'aresMedia',
+// };
+//
+// const multipleAresMetadataBlock = {
+//   model: {
+//     blocks: [
+//       {
+//         blockId: 'urn:bbc:ares::clip:p01k6msm',
+//         model: {
+//           advertising: true,
+//           caption: null,
+//           embedding: true,
+//           format: 'audio_video',
+//           id: 'p01k6msm',
+//           image: null,
+//           imageCopyright: 'BBC',
+//           imageUrl: 'ichef.test.bbci.co.uk/images/ic/$recipe/p01k6mtv.jpg',
+//           subType: 'clip',
+//           syndication: {
+//             destinations: [],
+//           },
+//           synopses: {
+//             short:
+//               'They may be tiny, but us humans could learn a thing or two from ants.',
+//           },
+//           title: 'Five things ants can teach us about management',
+//           versions: [
+//             {
+//               availableTerritories: {
+//                 nonUk: true,
+//                 uk: true,
+//               },
+//               availableUntil: null,
+//               duration: 191,
+//               types: ['Original'],
+//               versionId: 'p01k6msp',
+//               availableFrom: 1540218932,
+//               warnings: {
+//                 long: 'Contains strong language and adult humour.',
+//                 short: 'Contains strong language and adult humour.',
+//               },
+//             },
+//           ],
+//         },
+//         type: 'aresMediaMetadata',
+//       },
+//       {
+//         blockId: 'urn:bbc:ares::clip:p01k6mss',
+//         model: {
+//           advertising: true,
+//           caption: null,
+//           embedding: true,
+//           format: 'audio_video',
+//           id: 'p01k6mss',
+//           image: null,
+//           imageCopyright: 'BBC',
+//           imageUrl: 'ichef.test.bbci.co.uk/images/ic/$recipe/p01k6mtv.jpg',
+//           subType: 'clip',
+//           syndication: {
+//             destinations: [],
+//           },
+//           synopses: {
+//             short:
+//               'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor.',
+//           },
+//           title: 'Lorem ipsum is commonly used placeholder text.',
+//           versions: [
+//             {
+//               availableTerritories: {
+//                 nonUk: true,
+//                 uk: true,
+//               },
+//               availableUntil: null,
+//               duration: 191,
+//               types: ['Original'],
+//               versionId: 'p01k6msp',
+//               availableFrom: 1540218932,
+//               warnings: {
+//                 long: 'Contains common text.',
+//                 short: 'Contains common text.',
+//               },
+//             },
+//           ],
+//         },
+//         type: 'aresMediaMetadata',
+//       },
+//       imageBlock,
+//     ],
+//   },
+//   type: 'aresMedia',
+// };
+//
+// const validVideoWithCaptionBlocks = [captionBlock, validAresMediaVideoBlock];
+
+const missingVpidBlocks = [
+  captionBlock,
+  {
+    model: {
+      blocks: [
+        {
+          blockId: 'urn:bbc:ares::clip:p01k6msm',
+          model: {
+            advertising: true,
+            embedding: true,
+            format: 'audio_video',
+            id: 'p01k6msm',
+            imageCopyright: 'BBC',
+            imageUrl: 'ichef.test.bbci.co.uk/images/ic/$recipe/p01k6mtv.jpg',
+            subType: 'clip',
+            syndication: { destinations: [] },
+            synopses: {
+              short:
+                'They may be tiny, but us humans could learn a thing or two from ants.',
+            },
+            title: 'Five things ants can teach us about management',
+            versions: [
+              {
+                availableFrom: 1540218932000,
+                availableTerritories: { nonUk: true, uk: true },
+                duration: 191,
+                durationISO8601: 'PT3M11S',
+                types: ['Original'],
+                warnings: {
+                  long: 'Contains strong language and adult humour.',
+                  short: 'Contains strong language and adult humour.',
+                },
+              },
+            ],
+          },
+          type: 'aresMediaMetadata',
+          id: 'bede042c-ec9c-4462-8338-4b6fd9cde35d',
+          position: [4, 2, 1],
+        },
+        imageBlock,
+      ],
+    },
+    type: 'aresMedia',
+    id: 'e91c1a38-641d-4787-bec3-4f3783bb4b45',
+    position: [4, 2],
+  },
+];
+
+const defaultToggles = {
+  local: {
+    mediaPlayer: {
+      enabled: true,
+    },
+  },
+  test: {
+    mediaPlayer: {
+      enabled: true,
+    },
+  },
+  live: {
+    mediaPlayer: {
+      enabled: false,
+    },
+  },
+};
+
+const toggleStateOff = {
+  local: {
+    mediaPlayer: {
+      enabled: false,
+    },
+  },
+};
+
+const mockToggleDispatch = jest.fn();
+const GenerateFixtureData = ({
+  platform,
+  toggleState,
+  blocks,
+  placeholder,
+}) => (
+  <RequestContextProvider
+    isAmp={platform === 'amp'}
+    service="news"
+    statusCode={200}
+    platform={platform}
+    id="foo"
+    pageType="article"
+    pathname="/pathname"
+  >
+    <ServiceContextProvider service="news">
+      <ToggleContext.Provider
+        value={{ toggleState, toggleDispatch: mockToggleDispatch }}
+      >
+        <MediaPlayerContainer blocks={blocks} placeholder={placeholder} />
+      </ToggleContext.Provider>
+    </ServiceContextProvider>
+  </RequestContextProvider>
+);
+
+GenerateFixtureData.propTypes = {
+  platform: string.isRequired,
+  toggleState: shape({}),
+  blocks: oneOfType([object, arrayOf(object)]).isRequired,
+  placeholder: bool,
+};
+
+GenerateFixtureData.defaultProps = {
+  toggleState: defaultToggles,
+  placeholder: true,
+};
+
+export const VideoFixture = (
+  <GenerateFixtureData platform="canonical" blocks={validAresMediaVideoBlock} />
+);
+
+export const VideoAmpFixture = (
+  <GenerateFixtureData platform="amp" blocks={validAresMediaVideoBlock} />
+);
+
+export const VideoFixtureNoPlaceHolder = (
+  <GenerateFixtureData
+    platform="canonical"
+    blocks={validAresMediaVideoBlock}
+    placeholder={false}
+  />
+);
+
+export const VideoFixtureNoVersionId = (
+  <GenerateFixtureData platform="canonical" blocks={missingVpidBlocks} />
+);
+
+export const VideoFixtureToggledOff = (
+  <GenerateFixtureData
+    platform="canonical"
+    blocks={validAresMediaVideoBlock}
+    toggleState={toggleStateOff}
+  />
+);
